@@ -1,4 +1,5 @@
 import 'package:better_player/better_player.dart';
+import 'package:better_player/src/configuration/better_player_translations.dart';
 import 'package:better_player/src/controls/better_player_controls_configuration.dart';
 import 'package:better_player/src/subtitles/better_player_subtitles_configuration.dart';
 import 'package:flutter/material.dart';
@@ -16,9 +17,6 @@ class BetterPlayerConfiguration {
   /// Whether or not the video should loop
   final bool looping;
 
-  /// Weather or not to show the controls when initializing the widget.
-  final bool showControlsOnInitialize;
-
   /// When the video playback runs  into an error, you can build a custom
   /// error message.
   final Widget Function(BuildContext context, String errorMessage) errorBuilder;
@@ -33,6 +31,9 @@ class BetterPlayerConfiguration {
   /// or played.
   final Widget placeholder;
 
+  /// Should the placeholder be shown until play is pressed
+  final bool showPlaceholderUntilPlay;
+
   /// A widget which is placed between the video and the controls
   final Widget overlay;
 
@@ -41,6 +42,12 @@ class BetterPlayerConfiguration {
 
   /// Defines if the player will sleep in fullscreen or not
   final bool allowedScreenSleep;
+
+  /// Defines aspect ratio which will be used in fullscreen
+  final double fullScreenAspectRatio;
+
+  /// Defines the set of allowed device orientations on entering fullscreen
+  final List<DeviceOrientation> deviceOrientationsOnFullScreen;
 
   /// Defines the system overlays visible after exiting fullscreen
   final List<SystemUiOverlay> systemOverlaysAfterFullScreen;
@@ -71,6 +78,18 @@ class BetterPlayerConfiguration {
   ///Defines function which will react on player visibility changed
   final Function(double visibilityFraction) playerVisibilityChangedBehavior;
 
+  ///Defines translations used in player. If null, then default english translations
+  ///will be used.
+  final List<BetterPlayerTranslations> translations;
+
+  ///Defines if player should auto detect full screen device orientation based
+  ///on aspect ratio of the video. If aspect ratio of the video is < 1 then
+  ///video will played in full screen in portrait mode. If aspect ratio is >= 1
+  ///then video will be played horizontally. If this parameter is true, then
+  ///[deviceOrientationsOnFullScreen] and [fullScreenAspectRatio] value will be
+  /// ignored.
+  final bool autoDetectFullscreenDeviceOrientation;
+
   const BetterPlayerConfiguration({
     this.aspectRatio,
     this.autoPlay = false,
@@ -78,10 +97,15 @@ class BetterPlayerConfiguration {
     this.looping = false,
     this.fullScreenByDefault = false,
     this.placeholder,
+    this.showPlaceholderUntilPlay = false,
     this.overlay,
-    this.showControlsOnInitialize = true,
     this.errorBuilder,
     this.allowedScreenSleep = true,
+    this.fullScreenAspectRatio,
+    this.deviceOrientationsOnFullScreen = const [
+      DeviceOrientation.landscapeLeft,
+      DeviceOrientation.landscapeRight,
+    ],
     this.systemOverlaysAfterFullScreen = SystemUiOverlay.values,
     this.deviceOrientationsAfterFullScreen = const [
       DeviceOrientation.portraitUp,
@@ -96,6 +120,8 @@ class BetterPlayerConfiguration {
     this.fit = BoxFit.fill,
     this.rotation = 0,
     this.playerVisibilityChangedBehavior,
+    this.translations,
+    this.autoDetectFullscreenDeviceOrientation = false,
   });
 
   BetterPlayerConfiguration copyWith({
@@ -109,6 +135,8 @@ class BetterPlayerConfiguration {
     bool showControlsOnInitialize,
     Widget Function(BuildContext context, String errorMessage) errorBuilder,
     bool allowedScreenSleep,
+    double fullScreenAspectRatio,
+    List<DeviceOrientation> deviceOrientationsOnFullScreen,
     List<SystemUiOverlay> systemOverlaysAfterFullScreen,
     List<DeviceOrientation> deviceOrientationsAfterFullScreen,
     BetterPlayerRoutePageBuilder routePageBuilder,
@@ -118,32 +146,39 @@ class BetterPlayerConfiguration {
     BoxFit fit,
     double rotation,
     Function(double visibilityFraction) playerVisibilityChangedBehavior,
+    BetterPlayerTranslations translations,
+    bool autoDetectFullscreenDeviceOrientation,
   }) {
     return BetterPlayerConfiguration(
-        aspectRatio: aspectRatio ?? this.aspectRatio,
-        autoPlay: autoPlay ?? this.autoPlay,
-        startAt: startAt ?? this.startAt,
-        looping: looping ?? this.looping,
-        fullScreenByDefault: fullScreenByDefault ?? this.fullScreenByDefault,
-        placeholder: placeholder ?? this.placeholder,
-        overlay: overlay ?? this.overlay,
-        showControlsOnInitialize:
-            showControlsOnInitialize ?? this.showControlsOnInitialize,
-        errorBuilder: errorBuilder ?? this.errorBuilder,
-        allowedScreenSleep: allowedScreenSleep ?? this.allowedScreenSleep,
-        systemOverlaysAfterFullScreen:
-            systemOverlaysAfterFullScreen ?? this.systemOverlaysAfterFullScreen,
-        deviceOrientationsAfterFullScreen: deviceOrientationsAfterFullScreen ??
-            this.deviceOrientationsAfterFullScreen,
-        routePageBuilder: routePageBuilder ?? this.routePageBuilder,
-        eventListener: eventListener ?? this.eventListener,
-        subtitlesConfiguration:
-            subtitlesConfiguration ?? this.subtitlesConfiguration,
-        controlsConfiguration:
-            controlsConfiguration ?? this.controlsConfiguration,
-        fit: fit ?? this.fit,
-        rotation: rotation ?? this.rotation,
-        playerVisibilityChangedBehavior: playerVisibilityChangedBehavior ??
-            this.playerVisibilityChangedBehavior);
+      aspectRatio: aspectRatio ?? this.aspectRatio,
+      autoPlay: autoPlay ?? this.autoPlay,
+      startAt: startAt ?? this.startAt,
+      looping: looping ?? this.looping,
+      fullScreenByDefault: fullScreenByDefault ?? this.fullScreenByDefault,
+      placeholder: placeholder ?? this.placeholder,
+      overlay: overlay ?? this.overlay,
+      errorBuilder: errorBuilder ?? this.errorBuilder,
+      allowedScreenSleep: allowedScreenSleep ?? this.allowedScreenSleep,
+      deviceOrientationsOnFullScreen:
+          deviceOrientationsOnFullScreen ?? this.deviceOrientationsOnFullScreen,
+      systemOverlaysAfterFullScreen:
+          systemOverlaysAfterFullScreen ?? this.systemOverlaysAfterFullScreen,
+      deviceOrientationsAfterFullScreen: deviceOrientationsAfterFullScreen ??
+          this.deviceOrientationsAfterFullScreen,
+      routePageBuilder: routePageBuilder ?? this.routePageBuilder,
+      eventListener: eventListener ?? this.eventListener,
+      subtitlesConfiguration:
+          subtitlesConfiguration ?? this.subtitlesConfiguration,
+      controlsConfiguration:
+          controlsConfiguration ?? this.controlsConfiguration,
+      fit: fit ?? this.fit,
+      rotation: rotation ?? this.rotation,
+      playerVisibilityChangedBehavior: playerVisibilityChangedBehavior ??
+          this.playerVisibilityChangedBehavior,
+      translations: translations ?? this.translations,
+      autoDetectFullscreenDeviceOrientation:
+          autoDetectFullscreenDeviceOrientation ??
+              this.autoDetectFullscreenDeviceOrientation,
+    );
   }
 }
