@@ -1,7 +1,8 @@
 import 'dart:ui';
 
+import 'package:better_player/src/controls/better_player_overflow_menu_item.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:open_iconic_flutter/open_iconic_flutter.dart';
 
 class BetterPlayerControlsConfiguration {
   ///Color of the control bars
@@ -52,6 +53,9 @@ class BetterPlayerControlsConfiguration {
   ///Flag used to enable/disable play-pause
   final bool enablePlayPause;
 
+  ///Flag used to enable skip forward and skip back
+  final bool enableSkips;
+
   ///Progress bar played color
   final Color progressBarPlayedColor;
 
@@ -79,17 +83,12 @@ class BetterPlayerControlsConfiguration {
   ///Control bar height
   final double controlBarHeight;
 
-  ///Default error widget text
-  final String defaultErrorText;
-
-  ///Default loading next video text
-  final String loadingNextVideoText;
-
-  ///Text displayed when asset displayed in player is live stream
-  final String liveText;
-
   ///Live text color;
   final Color liveTextColor;
+
+  ///Flag used to show/hide overflow menu which contains playback, subtitles,
+  ///qualities options.
+  final bool enableOverflowMenu;
 
   ///Flag used to show/hide playback speed
   final bool enablePlaybackSpeed;
@@ -100,40 +99,66 @@ class BetterPlayerControlsConfiguration {
   ///Flag used to show/hide qualities
   final bool enableQualities;
 
-  const BetterPlayerControlsConfiguration({
-    this.controlBarColor = Colors.black87,
-    this.textColor = Colors.white,
-    this.iconsColor = Colors.white,
-    this.playIcon = Icons.play_arrow,
-    this.pauseIcon = Icons.pause,
-    this.muteIcon = Icons.volume_up,
-    this.unMuteIcon = Icons.volume_mute,
-    this.fullscreenEnableIcon = Icons.fullscreen,
-    this.fullscreenDisableIcon = Icons.fullscreen_exit,
-    this.skipBackIcon = OpenIconicIcons.reload,
-    this.skipForwardIcon = OpenIconicIcons.reload,
-    this.enableFullscreen = true,
-    this.enableMute = true,
-    this.enableProgressText = false,
-    this.enableProgressBar = true,
-    this.enablePlayPause = true,
-    this.progressBarPlayedColor = Colors.white,
-    this.progressBarHandleColor = Colors.white,
-    this.progressBarBufferedColor = Colors.white70,
-    this.progressBarBackgroundColor = Colors.white60,
-    this.controlsHideTime = const Duration(milliseconds: 300),
-    this.customControls,
-    this.showControls = true,
-    this.showControlsOnInitialize = true,
-    this.controlBarHeight = 48.0,
-    this.defaultErrorText = "Video can't be played",
-    this.loadingNextVideoText = "Loading next video",
-    this.liveText = "LIVE",
-    this.liveTextColor = Colors.red,
-    this.enablePlaybackSpeed = true,
-    this.enableSubtitles = true,
-    this.enableQualities = true,
-  });
+  ///Custom items of overflow menu
+  final List<BetterPlayerOverflowMenuItem> overflowMenuCustomItems;
+
+  ///Icon of the overflow menu
+  final IconData overflowMenuIcon;
+
+  ///Icon of the playback speed menu item from overflow menu
+  final IconData playbackSpeedIcon;
+
+  ///Icon of the subtitles menu item from overflow menu
+  final IconData subtitlesIcon;
+
+  ///Icon of the qualities menu item from overflow menu
+  final IconData qualitiesIcon;
+
+  ///Color of overflow menu icons
+  final Color overflowMenuIconsColor;
+
+  ///Time which will be used once user uses rewind and forward
+  final int skipsTimeInMilliseconds;
+
+  const BetterPlayerControlsConfiguration(
+      {this.controlBarColor = Colors.black87,
+      this.textColor = Colors.white,
+      this.iconsColor = Colors.white,
+      this.playIcon = Icons.play_arrow,
+      this.pauseIcon = Icons.pause,
+      this.muteIcon = Icons.volume_up,
+      this.unMuteIcon = Icons.volume_mute,
+      this.fullscreenEnableIcon = Icons.fullscreen,
+      this.fullscreenDisableIcon = Icons.fullscreen_exit,
+      this.skipBackIcon = Icons.fast_rewind,
+      this.skipForwardIcon = Icons.fast_forward,
+      this.enableFullscreen = true,
+      this.enableMute = true,
+      this.enableProgressText = false,
+      this.enableProgressBar = true,
+      this.enablePlayPause = true,
+      this.enableSkips = true,
+      this.progressBarPlayedColor = Colors.white,
+      this.progressBarHandleColor = Colors.white,
+      this.progressBarBufferedColor = Colors.white70,
+      this.progressBarBackgroundColor = Colors.white60,
+      this.controlsHideTime = const Duration(milliseconds: 300),
+      this.customControls,
+      this.showControls = true,
+      this.showControlsOnInitialize = true,
+      this.controlBarHeight = 48.0,
+      this.liveTextColor = Colors.red,
+      this.enableOverflowMenu = true,
+      this.enablePlaybackSpeed = true,
+      this.enableSubtitles = true,
+      this.enableQualities = true,
+      this.overflowMenuCustomItems = const [],
+      this.overflowMenuIcon = Icons.more_vert,
+      this.playbackSpeedIcon = Icons.shutter_speed,
+      this.qualitiesIcon = Icons.hd,
+      this.subtitlesIcon = Icons.text_fields,
+      this.overflowMenuIconsColor = Colors.black,
+      this.skipsTimeInMilliseconds = 15000});
 
   factory BetterPlayerControlsConfiguration.white() {
     return BetterPlayerControlsConfiguration(
@@ -148,10 +173,10 @@ class BetterPlayerControlsConfiguration {
 
   factory BetterPlayerControlsConfiguration.cupertino() {
     return BetterPlayerControlsConfiguration(
-        fullscreenEnableIcon: OpenIconicIcons.fullscreenEnter,
-        fullscreenDisableIcon: OpenIconicIcons.fullscreenExit,
-        playIcon: OpenIconicIcons.mediaPlay,
-        pauseIcon: OpenIconicIcons.mediaPause,
+        fullscreenEnableIcon: CupertinoIcons.fullscreen,
+        fullscreenDisableIcon: CupertinoIcons.fullscreen_exit,
+        playIcon: CupertinoIcons.play_arrow_solid,
+        pauseIcon: CupertinoIcons.pause_solid,
         enableProgressText: true);
   }
 }
